+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.6] - 2025-12-11
+
+### Changed
+- Indexing files in one level higher directories
+- Fix to read Windows paths - C\:/Users/user/OneDrive - Company... - from prop file
+
 ## [0.0.5] - 2025-12-10
 
 ### Added